covid_faq_assistant/requirements.txt
fastapi
uvicorn
streamlit
transformers
faiss-cpu
numpy
pandas
httpx
torch
